return version info accord setting

+++ src/consts.rs
@@ -39,8 +39,8 @@ pub const CONFLUX_TESTNET_CHAINID: usize = 1;
 /**
  * Application flags.
  */
-pub const APP_FLAG_BLIND_SIGNING_ENABLED: usize = 0x01;
+pub const APP_FLAG_BLIND_SIGNING_ENABLED: u8 = 0x01;
 
-pub const APP_FLAG_DETAILED_DISPLAY_ENABLED: usize = 0x02;
+pub const APP_FLAG_DETAILED_DISPLAY_ENABLED: u8 = 0x02;
 
 // pub const PERSONAL_SIGN_PREFIX: &str = "\x19Conflux Signed Message:\n";
